Import Blender text objects as labels via GLTF
The Blender GLTF exporter writes the generated text mesh of a text object to the node, needing special-case handling.

The Label implementation draws labels pointing towards +Z, which works for Collada models due to the implicit import rotation of the model.
Amend the orientation of label nodes when importing from GLTF.