SceneGraph: fold LOD node into Model
There is an implicit invariant that only one LOD node exists per model and it is at top-level. This change codifies that invariant, by moving the LOD node's responsibility into the Model itself.

A new algorithm is used to partition instance transforms. As a consequence, the amount of work done sorting instance transforms is cut in half, and memory use is drastically reduced compared to the previous implementation.